@@ -17,51 +17,39 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import { Row, Col } from 'components/graylog';
+import { Col } from 'components/graylog';
 import { IndicesConfiguration } from 'components/indices';
 
-const style = require('!style/useable!css!./IndexSetDetails.css');
+import StyledIndexSetDetailsRow from './StyledIndexSetDetailsRow';
 
-class IndexSetDetails extends React.Component {
-  static propTypes = {
-    indexSet: PropTypes.object.isRequired,
-  };
+const IndexSetDetails = ({ indexSet }) => {
+  return (
+    <StyledIndexSetDetailsRow>
+      <Col lg={3}>
+        <dl>
+          <dt>Index prefix:</dt>
+          <dd>{indexSet.index_prefix}</dd>
 
-  componentDidMount() {
-    style.use();
-  }
+          <dt>Shards:</dt>
+          <dd>{indexSet.shards}</dd>
 
-  componentWillUnmount() {
-    style.unuse();
-  }
+          <dt>Replicas:</dt>
+          <dd>{indexSet.replicas}</dd>
 
-  render() {
-    const { indexSet } = this.props;
+          <dt>Field type refresh interval:</dt>
+          <dd>{indexSet.field_type_refresh_interval / 1000.0} seconds</dd>
+        </dl>
+      </Col>
 
-    return (
-      <Row className="index-set-details">
-        <Col md={3}>
-          <dl>
-            <dt>Index prefix:</dt>
-            <dd>{indexSet.index_prefix}</dd>
+      <Col lg={6}>
+        <IndicesConfiguration indexSet={indexSet} />
+      </Col>
+    </StyledIndexSetDetailsRow>
+  );
+};
 
-            <dt>Shards:</dt>
-            <dd>{indexSet.shards}</dd>
-
-            <dt>Replicas:</dt>
-            <dd>{indexSet.replicas}</dd>
-
-            <dt>Field type refresh interval:</dt>
-            <dd>{indexSet.field_type_refresh_interval / 1000.0} seconds</dd>
-          </dl>
-        </Col>
-
-        <Col md={6}>
-          <IndicesConfiguration indexSet={indexSet} />
-        </Col>
-      </Row>
-    );
-  }
-}
+IndexSetDetails.propTypes = {
+  indexSet: PropTypes.object.isRequired,
+};
 
 export default IndexSetDetails;
