@@ -16,65 +16,102 @@
  */
 import PropTypes from 'prop-types';
 import React from 'react';
-import createReactClass from 'create-react-class';
-import Reflux from 'reflux';
 import lodash from 'lodash';
+import styled from 'styled-components';
 
 import { CounterDetails, GaugeDetails, HistogramDetails, MeterDetails, TimerDetails } from 'components/metrics';
 import CombinedProvider from 'injection/CombinedProvider';
+import connect from 'stores/connect';
 
 const { MetricsStore, MetricsActions } = CombinedProvider.get('Metrics');
-
-const MetricDetails = createReactClass({
-  displayName: 'MetricDetails',
-
-  propTypes: {
-    metric: PropTypes.object.isRequired,
-  },
-
-  mixins: [Reflux.connect(MetricsStore)],
-
-  componentDidMount() {
-    MetricsActions.add(this.props.nodeId, this.props.metric.full_name);
-  },
-
-  componentWillUnmount() {
-    MetricsActions.remove(this.props.nodeId, this.props.metric.full_name);
-  },
-
-  _formatDetailsForType(type, metric) {
-    switch (type) {
-      case 'Counter':
-        return <CounterDetails metric={metric} />;
-      case 'Gauge':
-        return <GaugeDetails metric={metric} />;
-      case 'Histogram':
-        return <HistogramDetails metric={metric} />;
-      case 'Meter':
-        return <MeterDetails metric={metric} />;
-      case 'Timer':
-        return <TimerDetails metric={metric} />;
-      default:
-        return <i>Invalid metric type: {type}</i>;
+const StyledMetricDetail = styled.div`
+  dl {
+    > dt {
+      float: left;
     }
-  },
 
-  render() {
-    const metricName = this.props.metric.full_name;
-    const { nodeId } = this.props;
-    const metric = this.state.metrics && this.state.metrics[nodeId] && this.state.metrics[nodeId][metricName]
-      ? this.state.metrics[nodeId][metricName] : this.props.metric;
-    const type = lodash.capitalize(metric.type);
-    const details = this._formatDetailsForType(type, metric);
+    &.metric-timer > dd {
+      margin-left: 145px;
+    }
 
-    return (
-      <div className="metric">
-        <h3>{type}</h3>
+    &.metric-meter > dd {
+      margin-left: 115px;
+    }
 
-        {details}
-      </div>
-    );
-  },
-});
+    &.metric-gauge > dd {
+      margin-left: 90px;
+    }
 
-export default MetricDetails;
+    &.metric-counter > dd {
+      margin-left: 90px;
+    }
+
+    &.metric-histogram > dd {
+      margin-left: 145px;
+    }
+  }
+`;
+
+class MetricDetails extends React.Component {
+ static propTypes = {
+   metrics: PropTypes.object,
+   metric: PropTypes.object.isRequired,
+   nodeId: PropTypes.string.isRequired,
+ };
+
+ static defaultProps = {
+   metrics: undefined,
+ }
+
+ componentDidMount() {
+   const { nodeId, metric: { full_name: fullName } } = this.props;
+   MetricsActions.add(nodeId, fullName);
+ }
+
+ componentWillUnmount() {
+   const { nodeId, metric: { full_name: fullName } } = this.props;
+   MetricsActions.remove(nodeId, fullName);
+ }
+
+ _formatDetailsForType = (type, metric) => {
+   switch (type) {
+     case 'Counter':
+       return <CounterDetails metric={metric} />;
+     case 'Gauge':
+       return <GaugeDetails metric={metric} />;
+     case 'Histogram':
+       return <HistogramDetails metric={metric} />;
+     case 'Meter':
+       return <MeterDetails metric={metric} />;
+     case 'Timer':
+       return <TimerDetails metric={metric} />;
+     default:
+       return <i>Invalid metric type: {type}</i>;
+   }
+ }
+
+ render() {
+   const { nodeId, metric, metric: { full_name: metricName }, metrics } = this.props;
+   const currentMetric = metrics?.[nodeId]?.[metricName] ?? metric;
+   const type = lodash.capitalize(currentMetric.type);
+   const details = this._formatDetailsForType(type, currentMetric);
+
+   return (
+     <div className="metric">
+       <h3>{type}</h3>
+       <StyledMetricDetail>
+         {details}
+       </StyledMetricDetail>
+     </div>
+   );
+ }
+}
+
+export default connect(
+  MetricDetails,
+  { metricsStore: MetricsStore },
+  ({ metricsStore, ...otherProps }) => ({
+    ...otherProps,
+    metrics: metricsStore.metrics,
+  }),
+);
