@@ -24,6 +24,7 @@ import View from 'views/logic/views/View';
 
 import Query from './Query';
 import WidgetGrid from './WidgetGrid';
+import WidgetFocusContext from './contexts/WidgetFocusContext';
 
 import AggregationWidget from '../logic/aggregationbuilder/AggregationWidget';
 import AggregationWidgetConfig from '../logic/aggregationbuilder/AggregationWidgetConfig';
@@ -48,6 +49,20 @@ const widget2 = AggregationWidget.builder()
 const widgets = Immutable.Map({ widget1, widget2 });
 
 describe('Query', () => {
+  const SUT = (props) => (
+    <WidgetFocusContext.Provider value={{ focusedWidget: undefined, setFocusedWidget: () => {} }}>
+      <Query results={{ errors: [], searchTypes: {} }}
+             widgetMapping={widgetMapping}
+             widgets={widgets}
+             onToggleMessages={() => {}}
+             queryId="someQueryId"
+             showMessages
+             allFields={Immutable.List()}
+             fields={Immutable.List()}
+             {...props} />
+    </WidgetFocusContext.Provider>
+  );
+
   it('renders extracted results and provided widgets', () => {
     const results = {
       errors: [],
@@ -57,14 +72,7 @@ describe('Query', () => {
       },
     };
     const wrapper = mount((
-      <Query results={results}
-             widgetMapping={widgetMapping}
-             widgets={widgets}
-             onToggleMessages={() => {}}
-             queryId="someQueryId"
-             showMessages
-             allFields={Immutable.List()}
-             fields={Immutable.List()} />
+      <SUT results={results} />
     ));
     const widgetGrid = wrapper.find(WidgetGrid);
 
@@ -83,14 +91,7 @@ describe('Query', () => {
       },
     };
     const wrapper = mount((
-      <Query results={results}
-             widgetMapping={widgetMapping}
-             widgets={widgets}
-             onToggleMessages={() => {}}
-             queryId="someQueryId"
-             showMessages
-             allFields={Immutable.List()}
-             fields={Immutable.List()} />
+      <SUT results={results} />
     ));
     const widgetGrid = wrapper.find(WidgetGrid);
 
@@ -110,14 +111,7 @@ describe('Query', () => {
       },
     };
     const wrapper = mount((
-      <Query results={results}
-             widgetMapping={widgetMapping}
-             widgets={widgets}
-             onToggleMessages={() => {}}
-             queryId="someQueryId"
-             showMessages
-             allFields={Immutable.List()}
-             fields={Immutable.List()} />
+      <SUT results={results} />
     ));
     const widgetGrid = wrapper.find(WidgetGrid);
 
@@ -135,14 +129,7 @@ describe('Query', () => {
       searchTypes: {},
     };
     const wrapper = mount((
-      <Query results={results}
-             widgetMapping={widgetMapping}
-             widgets={widgets}
-             onToggleMessages={() => {}}
-             queryId="someQueryId"
-             showMessages
-             allFields={Immutable.List()}
-             fields={Immutable.List()} />
+      <SUT results={results} />
     ));
     const widgetGrid = wrapper.find(WidgetGrid);
 
@@ -159,14 +146,8 @@ describe('Query', () => {
     };
     const wrapper = mount((
       <ViewTypeContext.Provider value={View.Type.Dashboard}>
-        <Query results={results}
-               widgetMapping={widgetMapping}
-               widgets={Immutable.Map()}
-               onToggleMessages={() => {}}
-               queryId="someQueryId"
-               showMessages
-               allFields={Immutable.List()}
-               fields={Immutable.List()} />
+        <SUT results={results}
+             widgets={Immutable.Map()} />
       </ViewTypeContext.Provider>
     ));
 
@@ -181,14 +162,8 @@ describe('Query', () => {
     };
     const wrapper = mount((
       <ViewTypeContext.Provider value={View.Type.Search}>
-        <Query results={results}
-               widgetMapping={widgetMapping}
-               widgets={Immutable.Map()}
-               onToggleMessages={() => {}}
-               queryId="someQueryId"
-               showMessages
-               allFields={Immutable.List()}
-               fields={Immutable.List()} />
+        <SUT results={results}
+             widgets={Immutable.Map()} />
       </ViewTypeContext.Provider>
     ));
 
@@ -202,14 +177,7 @@ describe('Query', () => {
       searchTypes: {},
     };
     const wrapper = mount((
-      <Query results={results}
-             widgetMapping={widgetMapping}
-             widgets={widgets}
-             onToggleMessages={() => {}}
-             queryId="someQueryId"
-             showMessages
-             allFields={Immutable.List()}
-             fields={Immutable.List()} />
+      <SUT results={results} />
     ));
 
     expect(wrapper.contains('You can create a new widget by selecting a widget type')).toEqual(false);
