@@ -14,7 +14,8 @@
  * along with this program. If not, see
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
-import React from 'react';
+import * as React from 'react';
+import { useContext } from 'react';
 import PropTypes from 'prop-types';
 import Immutable from 'immutable';
 import styled, { css } from 'styled-components';
@@ -22,13 +23,18 @@ import styled, { css } from 'styled-components';
 import DocsHelper from 'util/DocsHelper';
 import { Jumbotron } from 'components/graylog';
 import { CurrentViewStateActions } from 'views/stores/CurrentViewStateStore';
+import { useStore } from 'stores/connect';
 import { Spinner } from 'components/common';
 import { widgetDefinition } from 'views/logic/Widgets';
 import DocumentationLink from 'components/support/DocumentationLink';
 import IfDashboard from 'views/components/dashboard/IfDashboard';
 import IfSearch from 'views/components/search/IfSearch';
-import WidgetGrid from 'views/components/WidgetGrid';
+import WidgetGrid, { WidgetContainer } from 'views/components/WidgetGrid';
 import WidgetPosition from 'views/logic/widgets/WidgetPosition';
+import WidgetComponent from 'views/components/WidgetComponent';
+import { TitlesStore, TitleTypes } from 'views/stores/TitlesStore';
+import defaultTitle from 'views/components/defaultTitle';
+import WidgetFocusContext from 'views/components/contexts/WidgetFocusContext';
 
 import { PositionsMap, ImmutableWidgetsMap } from './widgets/WidgetPropTypes';
 import InteractiveContext from './contexts/InteractiveContext';
@@ -52,33 +58,70 @@ const _onPositionsChange = (positions) => {
   CurrentViewStateActions.widgetPositions(newPositions);
 };
 
+const _getDataAndErrors = (widget, widgetMapping, results) => {
+  const { searchTypes } = results;
+  const widgetType = widgetDefinition(widget.type);
+  const dataTransformer = widgetType.searchResultTransformer || ((x) => x);
+  const searchTypeIds = (widgetMapping[widget.id] || []);
+  const widgetData = searchTypeIds.map((searchTypeId) => searchTypes[searchTypeId]).filter((result) => result);
+  const widgetErrors = results.errors.filter((e) => searchTypeIds.includes(e.searchTypeId));
+  let error;
+
+  const data = dataTransformer(widgetData, widget);
+
+  if (widgetErrors && widgetErrors.length > 0) {
+    error = widgetErrors;
+  }
+
+  if (!widgetData || widgetData.length === 0) {
+    const queryErrors = results.errors.filter((e) => e.type === 'query');
+
+    if (queryErrors.length > 0) {
+      error = error ? [].concat(error, queryErrors) : queryErrors;
+    }
+  }
+
+  return { widgetData: data, error };
+};
+
+const _renderFocusedWidget = (focusedWidget, widgetDefs, titles, widgetMapping, fields, results) => {
+  const widget = widgetDefs.find((w) => w.id === focusedWidget);
+
+  if (widget) {
+    const title = titles.getIn([TitleTypes.Widget, widget.id], defaultTitle(widget));
+    const { widgetData, error } = _getDataAndErrors(widget, widgetMapping, results);
+    const data = { [widget.id]: widgetData };
+    const errors = { [widget.id]: error };
+
+    return (
+      <WidgetContainer>
+        <WidgetComponent widget={widget}
+                         widgetId={widget.id}
+                         data={data}
+                         errors={errors}
+                         widgetDimension={{ height: 100, width: 200 }}
+                         title={title}
+                         position={WidgetPosition.builder().build()}
+                         fields={fields} />
+      </WidgetContainer>
+    );
+  }
+
+  return undefined;
+};
+
 const _renderWidgetGrid = (widgetDefs, widgetMapping, results, positions, queryId, fields, allFields) => {
   const widgets = {};
   const data = {};
   const errors = {};
-  const { searchTypes } = results;
 
   widgetDefs.forEach((widget) => {
-    const widgetType = widgetDefinition(widget.type);
-    const dataTransformer = widgetType.searchResultTransformer || ((x) => x);
-    const searchTypeIds = (widgetMapping[widget.id] || []);
-    const widgetData = searchTypeIds.map((searchTypeId) => searchTypes[searchTypeId]).filter((result) => result);
-    const widgetErrors = results.errors.filter((e) => searchTypeIds.includes(e.searchTypeId));
-
     widgets[widget.id] = widget;
-    data[widget.id] = dataTransformer(widgetData, widget);
 
-    if (widgetErrors && widgetErrors.length > 0) {
-      errors[widget.id] = widgetErrors;
-    }
+    const { widgetData, error } = _getDataAndErrors(widget, widgetMapping, results);
 
-    if (!widgetData || widgetData.length === 0) {
-      const queryErrors = results.errors.filter((e) => e.type === 'query');
-
-      if (queryErrors.length > 0) {
-        errors[widget.id] = errors[widget.id] ? [].concat(errors[widget.id], queryErrors) : queryErrors;
-      }
-    }
+    data[widget.id] = widgetData;
+    errors[widget.id] = error;
   });
 
   return (
@@ -129,14 +172,19 @@ const EmptyDashboardInfo = () => (
 );
 
 const Query = ({ allFields, fields, results, positions, widgetMapping, widgets, queryId }) => {
+  const { focusedWidget } = useContext(WidgetFocusContext);
+  const titles = useStore(TitlesStore);
+
   if (!widgets || widgets.isEmpty()) {
     return <EmptyDashboardInfo />;
   }
 
   if (results) {
-    const content = _renderWidgetGrid(widgets, widgetMapping.toJS(), results, positions, queryId, fields, allFields);
+    const content = focusedWidget
+      ? _renderFocusedWidget(focusedWidget, widgets, titles, widgetMapping.toJS(), fields, results)
+      : _renderWidgetGrid(widgets, widgetMapping.toJS(), results, positions, queryId, fields, allFields);
 
-    return (<span>{content}</span>);
+    return (<>{content}</>);
   }
 
   return <Spinner />;
