@@ -36,6 +36,7 @@ import CurrentViewTypeProvider from 'views/components/views/CurrentViewTypeProvi
 import ViewTypeContext from 'views/components/contexts/ViewTypeContext';
 
 import Search from './Search';
+import DashboardSearchBar from './DashboardSearchBar';
 
 import { useSyncWithQueryParameters } from '../hooks/SyncWithQueryParameters';
 
@@ -166,23 +167,6 @@ describe('Search', () => {
     expect(SearchConfigActions.refresh).toHaveBeenCalled();
   });
 
-  it('does not register to WidgetStore upon mount', () => {
-    mount(<SimpleSearch />);
-
-    expect(WidgetStore.listen).not.toHaveBeenCalled();
-  });
-
-  it('does not unregister from Widget store upon unmount', () => {
-    const unsubscribe = jest.fn();
-
-    WidgetStore.listen = jest.fn(() => unsubscribe);
-    const wrapper = mount(<SimpleSearch />);
-
-    wrapper.unmount();
-
-    expect(unsubscribe).not.toHaveBeenCalled();
-  });
-
   it('does not register to QueryFiltersStore upon mount', () => {
     mount(<SimpleSearch />);
 
@@ -281,12 +265,12 @@ describe('Search', () => {
   it('refreshing after query change parses search metadata first', (done) => {
     const wrapper = mount(<SimpleSearch />);
 
-    const searchBar = wrapper.find('DashboardSearchBar');
-    const cb = searchBar.at(0).props().onExecute;
+    const searchBar = wrapper.find(DashboardSearchBar);
+    const cb = searchBar.at(0).props().onExecute as (view?: View) => Promise<void>;
 
     const view = { search: {} };
 
-    const promise = cb(view);
+    const promise = cb(view as View);
 
     promise.then(() => {
       expect(SearchMetadataActions.parseSearch).toHaveBeenCalled();
