@@ -15,16 +15,18 @@
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
 import * as React from 'react';
+import { useContext, useState, useCallback } from 'react';
 import styled, { css } from 'styled-components';
-import { PluginStore } from 'graylog-web-plugin/plugin';
 
 import { MenuItem } from 'components/graylog';
 import FieldType from 'views/logic/fieldtypes/FieldType';
 import { ActionContext } from 'views/logic/ActionContext';
 import type { QueryId } from 'views/logic/queries/Query';
+import WidgetFocusContext from 'views/components/contexts/WidgetFocusContext';
+import usePluginEntities from 'views/logic/usePluginEntities';
 
+import type { ActionDefinition } from './ActionHandler';
 import { createHandlerFor } from './ActionHandler';
-import type { ActionComponents, ActionDefinition } from './ActionHandler';
 
 import OverlayDropdown from '../OverlayDropdown';
 import style from '../Field.css';
@@ -39,11 +41,6 @@ type Props = {
   type: FieldType,
 };
 
-type State = {
-  open: boolean,
-  overflowingComponents: ActionComponents,
-};
-
 type FieldElementProps = {
   active: boolean,
   disabled: boolean,
@@ -56,71 +53,71 @@ const FieldElement = styled.span.attrs({
   opacity: ${disabled ? '0.3' : '1'};
 `);
 
-class FieldActions extends React.Component<Props, State> {
-  static contextType = ActionContext;
+const FieldActions = ({ children, disabled, element, menuContainer, name, type, queryId }: Props) => {
+  const actionContext = useContext(ActionContext);
+  const { setFocusedWidget } = useContext(WidgetFocusContext);
+  const allFieldActions = usePluginEntities('fieldActions');
 
-  constructor(props: Props, context: typeof ActionContext) {
-    super(props, context);
+  const [open, setOpen] = useState(false);
+  const [overflowingComponents, setOverflowingComponents] = useState({});
 
-    this.state = {
-      open: false,
-      overflowingComponents: {},
-    };
-  }
+  const _onMenuToggle = useCallback(() => setOpen(!open), [open]);
 
-  _onMenuToggle = () => this.setState((state) => ({ open: !state.open }));
+  const wrappedElement = <FieldElement active={open} disabled={disabled}>{element}</FieldElement>;
+  const handlerArgs = { queryId, field: name, type, contexts: actionContext };
+  const fieldActions = allFieldActions
+    .filter((action: ActionDefinition) => {
+      const { isHidden = () => false } = action;
 
-  render() {
-    const { children, disabled, element, menuContainer, name, type, queryId } = this.props;
-    const { open } = this.state;
+      return !isHidden(handlerArgs);
+    })
+    .map((action: ActionDefinition) => {
+      const setActionComponents = (fn) => {
+        setOverflowingComponents(fn(overflowingComponents));
+      };
 
-    const wrappedElement = <FieldElement active={open} disabled={disabled}>{element}</FieldElement>;
-    const handlerArgs = { queryId, field: name, type, contexts: this.context };
-    const fieldActions: Array<ActionDefinition> = PluginStore.exports('fieldActions')
-      .filter((action: ActionDefinition) => {
-        const { isHidden = () => false } = action;
+      const handler = createHandlerFor(action, setActionComponents);
 
-        return !isHidden(handlerArgs);
-      })
-      .map((action: ActionDefinition) => {
-        const setActionComponents = (fn) => this.setState(({ overflowingComponents: actionComponents }) => ({ overflowingComponents: fn(actionComponents) }));
-        const handler = createHandlerFor(action, setActionComponents);
+      const onSelect = () => {
+        const { resetFocus = false } = action;
 
-        const onSelect = () => {
-          this._onMenuToggle();
-          handler(handlerArgs);
-        };
+        if (resetFocus) {
+          setFocusedWidget(undefined);
+        }
 
-        const { isEnabled = () => true } = action;
-        const actionDisabled = !isEnabled(handlerArgs);
+        _onMenuToggle();
+        handler(handlerArgs);
+      };
 
-        return (
-          <MenuItem key={`${name}-action-${action.type}`}
-                    disabled={actionDisabled}
-                    eventKey={{ action: action.type, field: name }}
-                    onSelect={onSelect}>{action.title}
-          </MenuItem>
-        );
-      });
+      const { isEnabled = () => true } = action;
+      const actionDisabled = !isEnabled(handlerArgs);
 
-    return (
-      <OverlayDropdown show={open}
-                       toggle={wrappedElement}
-                       placement="right"
-                       onToggle={this._onMenuToggle}
-                       menuContainer={menuContainer}>
-        <div style={{ marginBottom: '10px' }}>
-          <span className={`field-name ${style.dropdownheader}`}>
-            {children}
-          </span>
-        </div>
+      return (
+        <MenuItem key={`${name}-action-${action.type}`}
+                  disabled={actionDisabled}
+                  eventKey={{ action: action.type, field: name }}
+                  onSelect={onSelect}>{action.title}
+        </MenuItem>
+      );
+    });
 
-        <MenuItem divider />
-        <MenuItem header>Actions</MenuItem>
-        {fieldActions}
-      </OverlayDropdown>
-    );
-  }
-}
+  return (
+    <OverlayDropdown show={open}
+                     toggle={wrappedElement}
+                     placement="right"
+                     onToggle={_onMenuToggle}
+                     menuContainer={menuContainer}>
+      <div style={{ marginBottom: '10px' }}>
+        <span className={`field-name ${style.dropdownheader}`}>
+          {children}
+        </span>
+      </div>
+
+      <MenuItem divider />
+      <MenuItem header>Actions</MenuItem>
+      {fieldActions}
+    </OverlayDropdown>
+  );
+};
 
 export default FieldActions;
