@@ -59,14 +59,18 @@ const createViewWithQueries = () => {
 };
 
 describe('BigDisplayModeConfiguration', () => {
+  const SUT = (props) => (
+    <BigDisplayModeConfiguration view={view} {...props} />
+  );
+
   it('generates markup that matches snapshot', () => {
-    const { container } = render(<BigDisplayModeConfiguration view={view} />);
+    const { container } = render(<SUT />);
 
     expect(container).toMatchSnapshot();
   });
 
   it('disables menu item if `disabled` prop is `true`', () => {
-    const { getByText, queryByText } = render(<BigDisplayModeConfiguration view={view} disabled />);
+    const { getByText, queryByText } = render(<SUT disabled />);
     const menuItem = getByText('Full Screen');
 
     fireEvent.submit(menuItem);
@@ -75,7 +79,7 @@ describe('BigDisplayModeConfiguration', () => {
   });
 
   it('opens modal when menu item is clicked', async () => {
-    const { findByText, getByText } = render(<BigDisplayModeConfiguration view={view} />);
+    const { findByText, getByText } = render(<SUT />);
     const menuItem = getByText('Full Screen');
 
     fireEvent.click(menuItem);
@@ -84,14 +88,14 @@ describe('BigDisplayModeConfiguration', () => {
   });
 
   it('shows open modal per default if `open` prop is `true`', () => {
-    const { getByText } = render(<BigDisplayModeConfiguration view={view} show />);
+    const { getByText } = render(<SUT show />);
 
     expect(getByText('Configuring Full Screen')).not.toBeNull();
   });
 
   it('shows all query titles in modal', () => {
     const viewWithQueries = createViewWithQueries();
-    const { getByText } = render(<BigDisplayModeConfiguration view={viewWithQueries} show />);
+    const { getByText } = render(<SUT view={viewWithQueries} show />);
 
     expect(getByText('Page#1')).not.toBeNull();
     expect(getByText('My awesome Query tab')).not.toBeNull();
@@ -99,7 +103,7 @@ describe('BigDisplayModeConfiguration', () => {
   });
 
   it('should not allow strings for the refresh interval', () => {
-    const { getByLabelText } = render(<BigDisplayModeConfiguration view={view} show />);
+    const { getByLabelText } = render(<SUT show />);
 
     const refreshInterval = asElement(getByLabelText('Refresh Interval'), HTMLInputElement);
 
@@ -109,7 +113,7 @@ describe('BigDisplayModeConfiguration', () => {
   });
 
   it('should not allow strings for the cycle interval', () => {
-    const { getByLabelText } = render(<BigDisplayModeConfiguration view={view} show />);
+    const { getByLabelText } = render(<SUT show />);
 
     const cycleInterval = asElement(getByLabelText('Tab cycle interval'), HTMLInputElement);
 
@@ -125,7 +129,7 @@ describe('BigDisplayModeConfiguration', () => {
     });
 
     it('on form submit', () => {
-      const { getByTestId } = render(<BigDisplayModeConfiguration view={view} show />);
+      const { getByTestId } = render(<SUT show />);
       const form = getByTestId('modal-form');
 
       expect(form).not.toBeNull();
@@ -137,7 +141,7 @@ describe('BigDisplayModeConfiguration', () => {
     });
 
     it('including changed refresh interval', () => {
-      const { getByLabelText, getByTestId } = render(<BigDisplayModeConfiguration view={view} show />);
+      const { getByLabelText, getByTestId } = render(<SUT show />);
 
       const refreshInterval = getByLabelText('Refresh Interval');
 
@@ -152,7 +156,7 @@ describe('BigDisplayModeConfiguration', () => {
     });
 
     it('including tab cycle interval setting', () => {
-      const { getByLabelText, getByTestId } = render(<BigDisplayModeConfiguration view={view} show />);
+      const { getByLabelText, getByTestId } = render(<SUT show />);
 
       const cycleInterval = getByLabelText('Tab cycle interval');
 
@@ -168,7 +172,7 @@ describe('BigDisplayModeConfiguration', () => {
 
     it('including selected tabs', () => {
       const viewWithQueries = createViewWithQueries();
-      const { getByLabelText, getByTestId } = render(<BigDisplayModeConfiguration view={viewWithQueries} show />);
+      const { getByLabelText, getByTestId } = render(<SUT view={viewWithQueries} show />);
 
       const query1 = getByLabelText('Page#1');
 
