@@ -14,7 +14,7 @@
  * along with this program. If not, see
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
-import React, { useCallback, useState } from 'react';
+import React, { useCallback, useState, useContext } from 'react';
 import URI from 'urijs';
 
 import history from 'util/History';
@@ -25,6 +25,7 @@ import { Icon } from 'components/common';
 import Routes from 'routing/Routes';
 import View from 'views/logic/views/View';
 import queryTitle from 'views/logic/queries/QueryTitle';
+import WidgetFocusContext from 'views/components/contexts/WidgetFocusContext';
 
 export type UntypedBigDisplayModeQuery = {
   interval?: string,
@@ -112,11 +113,15 @@ const ConfigurationModal = ({ onSave, view, show, onClose }: ConfigurationModalP
   );
 };
 
-const redirectToBigDisplayMode = (view: View, config: UntypedBigDisplayModeQuery): void => history.push(
-  new URI(Routes.pluginRoute('DASHBOARDS_TV_VIEWID')(view.id))
-    .search(config)
-    .toString(),
-);
+const redirectToBigDisplayMode = (view: View, config: UntypedBigDisplayModeQuery, setFocusedWidget): void => {
+  setFocusedWidget(undefined);
+
+  history.push(
+    new URI(Routes.pluginRoute('DASHBOARDS_TV_VIEWID')(view.id))
+      .search(config)
+      .toString(),
+  );
+};
 
 const createQueryFromConfiguration = (
   { queryCycleInterval, queryTabs, refreshInterval }: Configuration,
@@ -141,7 +146,8 @@ type Props = {
 
 const BigDisplayModeConfiguration = ({ disabled, show, view }: Props) => {
   const [showConfigurationModal, setShowConfigurationModal] = useState(show);
-  const onSave = (config: Configuration) => redirectToBigDisplayMode(view, createQueryFromConfiguration(config, view));
+  const { setFocusedWidget } = useContext(WidgetFocusContext);
+  const onSave = (config: Configuration) => redirectToBigDisplayMode(view, createQueryFromConfiguration(config, view), setFocusedWidget);
 
   return (
     <>
