@@ -17,6 +17,7 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
 import { last, sortBy } from 'lodash';
+import { withTheme, DefaultTheme } from 'styled-components';
 
 import StringUtils from 'util/StringUtils';
 import { DEFAULT_HIGHLIGHT_COLOR } from 'views/Constants';
@@ -39,6 +40,7 @@ type Props = {
   field: string,
   value?: any,
   highlightRanges: Ranges,
+  theme: DefaultTheme,
 };
 
 function highlightCompleteValue(ranges: Array<HighlightRange>, value) {
@@ -54,8 +56,8 @@ function highlightCompleteValue(ranges: Array<HighlightRange>, value) {
 
 const shouldBeFormatted = (field, value) => isFunction(field) && isNumeric(value);
 
-const PossiblyHighlight = ({ color = DEFAULT_HIGHLIGHT_COLOR, field, value, highlightRanges = {} }: Props) => {
-  if (value === undefined || value == null) {
+const PossiblyHighlight = ({ color = DEFAULT_HIGHLIGHT_COLOR, field, value, highlightRanges = {}, theme }: Props) => {
+  if (value === undefined || value === null) {
     return '';
   }
 
@@ -67,6 +69,8 @@ const PossiblyHighlight = ({ color = DEFAULT_HIGHLIGHT_COLOR, field, value, high
 
   const style = {
     backgroundColor: color,
+    color: theme.utils.contrastingColor(color),
+    padding: '0 1px',
   };
 
   if (highlightCompleteValue(highlightRanges[field], value)) {
@@ -115,4 +119,4 @@ PossiblyHighlight.defaultProps = {
   value: undefined,
 };
 
-export default PossiblyHighlight;
+export default withTheme(PossiblyHighlight);
