@@ -15,7 +15,7 @@
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
 import * as React from 'react';
-import { act, fireEvent, render, screen } from 'wrappedTestingLibrary';
+import { render } from 'wrappedTestingLibrary';
 import { Formik, Form } from 'formik';
 
 import AbsoluteRangeField from './AbsoluteRangeField';
@@ -54,87 +54,5 @@ describe('AbsoluteRangeField', () => {
     ));
 
     expect(asFragment()).toMatchSnapshot();
-  });
-
-  it('toggles bod & eod', () => {
-    renderWithForm((
-      <AbsoluteRangeField {...defaultProps} from />
-    ));
-
-    const toggleBtn = screen.getByRole('button', { name: /toggle between beginning and end of day/i });
-    fireEvent.click(toggleBtn);
-
-    const inputHour = screen.getByRole('spinbutton', { name: /from hour/i });
-    const inputMinutes = screen.getByRole('spinbutton', { name: /from minutes/i });
-    const inputSeconds = screen.getByRole('spinbutton', { name: /from seconds/i });
-    const inputMillsecs = screen.getByRole('spinbutton', { name: /from milliseconds/i });
-
-    expect(inputHour).toHaveValue(0);
-    expect(inputMinutes).toHaveValue(0);
-    expect(inputSeconds).toHaveValue(0);
-    expect(inputMillsecs).toHaveValue(0);
-
-    fireEvent.click(toggleBtn);
-
-    expect(inputHour).toHaveValue(23);
-    expect(inputMinutes).toHaveValue(59);
-    expect(inputSeconds).toHaveValue(59);
-    expect(inputMillsecs).toHaveValue(999);
-  });
-
-  it('does not allow non-numeric characters', () => {
-    renderWithForm((
-      <AbsoluteRangeField {...defaultProps} from />
-    ));
-
-    const inputHour = screen.getByRole('spinbutton', { name: /from hour/i });
-
-    act(() => {
-      fireEvent.change(inputHour, { target: { value: '/w!' } });
-    });
-
-    expect(inputHour).toHaveValue(0);
-  });
-
-  it('does allow proper value', () => {
-    renderWithForm((
-      <AbsoluteRangeField {...defaultProps} from />
-    ));
-
-    const inputHour = screen.getByRole('spinbutton', { name: /from hour/i });
-
-    act(() => {
-      fireEvent.change(inputHour, { target: { value: '10' } });
-    });
-
-    expect(inputHour).toHaveValue(10);
-  });
-
-  it('does not allow numbers over their maximum', () => {
-    renderWithForm((
-      <AbsoluteRangeField {...defaultProps} from />
-    ));
-
-    const inputHour = screen.getByRole('spinbutton', { name: /from hour/i });
-
-    act(() => {
-      fireEvent.change(inputHour, { target: { value: '50' } });
-    });
-
-    expect(inputHour).toHaveValue(23);
-  });
-
-  it('does not try to parse an empty date', () => {
-    renderWithForm((
-      <AbsoluteRangeField {...defaultProps} from />
-    ));
-
-    const inputHour = screen.getByRole('spinbutton', { name: /from hour/i });
-
-    act(() => {
-      fireEvent.change(inputHour, { target: { value: '' } });
-    });
-
-    expect(inputHour).toHaveValue(0);
   });
 });
