@@ -108,17 +108,17 @@ const DEFAULT_RANGES = {
 };
 
 const timeRangeTypeTabs = ({ activeTab, originalTimeRange, limitDuration, currentTimeRange }) => availableTimeRangeTypes.map(({ type, name }) => {
-  const RangeComponent = timeRangeTypes?.[type] || DisabledTimeRangeSelector;
+  const TimeRangeComponent = timeRangeTypes[type];
 
   return (
     <Tab title={name}
          key={`time-range-type-selector-${type}`}
          eventKey={type}>
       {type === activeTab && (
-        <RangeComponent disabled={false}
-                        originalTimeRange={originalTimeRange || DEFAULT_RANGES[type]}
-                        limitDuration={limitDuration}
-                        currentTimeRange={currentTimeRange || DEFAULT_RANGES[type]} />
+        <TimeRangeComponent disabled={false}
+                            originalTimeRange={originalTimeRange || DEFAULT_RANGES[type]}
+                            limitDuration={limitDuration}
+                            currentTimeRange={currentTimeRange || DEFAULT_RANGES[type]} />
       )}
     </Tab>
   );
@@ -195,7 +195,7 @@ const TimeRangeDropdown = ({ noOverride, toggleDropdownShow }: Props) => {
 
           <StyledTabs id="dateTimeTypes"
                       defaultActiveKey={availableTimeRangeTypes[0].type}
-                      activeKey={activeTab}
+                      activeKey={activeTab ?? -1}
                       onSelect={setActiveTab}
                       animation={false}>
             {timeRangeTypeTabs({
@@ -204,6 +204,9 @@ const TimeRangeDropdown = ({ noOverride, toggleDropdownShow }: Props) => {
               limitDuration,
               currentTimeRange,
             })}
+
+            {!activeTab && (<DisabledTimeRangeSelector />)}
+
           </StyledTabs>
         </Col>
       </Row>
