@@ -136,20 +136,24 @@ describe('MessageList', () => {
 
     SelectedFieldsStore.getInitialState = jest.fn(() => Immutable.Set([TIMESTAMP_FIELD, 'file_name']));
     const config = MessagesWidgetConfig.builder().fields([TIMESTAMP_FIELD, 'file_name']).build();
-    const wrapper1 = mount(<MessageList data={data}
-                                        config={config}
-                                        fields={Immutable.List(fields)}
-                                        setLoadingState={() => {}} />);
+    const wrapper1 = mount(
+      <MessageList data={data}
+                   config={config}
+                   fields={Immutable.List(fields)}
+                   setLoadingState={() => {}} />,
+    );
 
     expect(wrapper1.find('span[role="presentation"]').length).toBe(2);
 
     const emptyConfig = MessagesWidgetConfig.builder().fields([]).build();
 
     SelectedFieldsStore.getInitialState = jest.fn(() => Immutable.Set([]));
-    const wrapper2 = mount(<MessageList data={data}
-                                        config={emptyConfig}
-                                        fields={Immutable.List(fields)}
-                                        setLoadingState={() => {}} />);
+    const wrapper2 = mount(
+      <MessageList data={data}
+                   config={emptyConfig}
+                   fields={Immutable.List(fields)}
+                   setLoadingState={() => {}} />,
+    );
 
     expect(wrapper2.find('span[role="presentation"]').length).toBe(0);
   });
@@ -157,10 +161,12 @@ describe('MessageList', () => {
   it('provides a message context for each individual entry', () => {
     const fields = [new FieldTypeMapping('file_name', new FieldType('string', ['full-text-search'], []))];
     const config = MessagesWidgetConfig.builder().fields(['file_name']).build();
-    const wrapper = mount(<MessageList data={data}
-                                       fields={Immutable.List(fields)}
-                                       config={config}
-                                       setLoadingState={() => {}} />);
+    const wrapper = mount(
+      <MessageList data={data}
+                   fields={Immutable.List(fields)}
+                   config={config}
+                   setLoadingState={() => {}} />,
+    );
     const messageTableEntry = wrapper.find('MessageTableEntry');
     const td = messageTableEntry.find('td').at(0);
 
