@@ -156,9 +156,9 @@ export const ViewStore: ViewStoreType = singletonStore(
 
       /* Select selected query (activeQuery) or first query in view (for now).
          Selected query might become a property on the view later. */
-      const queries = get(view, 'search.queries', Immutable.List());
-      const firstQueryId = get(queries.first(), 'id');
-      const selectedQuery = this.activeQuery && queries.find((q) => (q.id === this.activeQuery)) ? this.activeQuery : firstQueryId;
+      const queries = view.state.keySeq().toList();
+      const firstQueryId = queries.first();
+      const selectedQuery = this.activeQuery && queries.find((id) => (id === this.activeQuery)) ? this.activeQuery : firstQueryId;
 
       this.selectQuery(selectedQuery);
       this.isNew = isNew;
